Correction adresse_internet, socket_tcp.

+++ socket_tcp/socket_tcp.h
@@ -27,10 +27,11 @@ extern int creerSocketEcouteTCP(SocketTCP *isocket, const char *adresse,
 
 extern int acceptSocketTCP(const SocketTCP *secoute, SocketTCP *sservice);
 
-extern int writeSocketTCP(const SocketTCP *osocket, const void *buffer,
-                          size_t length);
+extern ssize_t writeSocketTCP(const SocketTCP *osocket, const void *buffer,
+                              size_t length);
 
-extern int readSocketTCP(const SocketTCP *nsocket, void *buffer, size_t length);
+extern ssize_t readSocketTCP(const SocketTCP *nsocket, void *buffer,
+                             size_t length);
 
 extern int closeSocketTCP(SocketTCP *socket);
